dom/animation/test/mozilla/test_get_animations_on_scroll_animations.html

<!doctype html>
<head>
<meta charset=utf-8>
<title>Test getAnimations() which doesn't return scroll animations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<style>
  @keyframes animWidth {
    from { width: 100px; }
    to { width: 200px }
  }
  @keyframes animTop {
    to { top: 100px }
  }
  .fill-vh {
    width: 100px;
    height: 100vh;
  }
</style>
</head>
<body>
<div id="log"></div>
<script>
"use strict";

test(function(t) {
  const div = addDiv(t,
    { style: "width: 10px; height: 100px; " +
             "animation: animWidth 100s scroll(), animTop 200s;" });

  // Sanity check to make sure the scroll animation is there.
  addDiv(t, { class: "fill-vh" });
  const scroller = document.scrollingElement;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;
  scroller.scrollTop = maxScroll;
  assert_equals(getComputedStyle(div).width, "200px",
                "The scroll animation is there");

  const animations = div.getAnimations();
  assert_equals(animations.length, 2,
                'getAnimations() should include scroll animations');
  assert_equals(animations[0].animationName, "animWidth",
                'getAmimations() should return scroll animations');
  // FIXME: Bug 1676794. Support ScrollTimeline interface.
  assert_equals(animations[0].timeline, null,
                'scroll animation should not return scroll timeline');
}, 'Element.getAnimation() should include scroll animations');

test(function(t) {
  const div = addDiv(t,
    { style: "width: 10px; height: 100px; " +
             "animation: animWidth 100s scroll(), animTop 100s;" });

  // Sanity check to make sure the scroll animation is there.
  addDiv(t, { class: "fill-vh" });
  const scroller = document.scrollingElement;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;
  scroller.scrollTop = maxScroll;
  assert_equals(getComputedStyle(div).width, "200px",
                "The scroll animation is there");

  const animations = document.getAnimations();
  assert_equals(animations.length, 2,
                'getAnimations() should include scroll animations');
  assert_equals(animations[0].animationName, "animWidth",
                'getAmimations() should return scroll animations');
  // FIXME: Bug 1676794. Support ScrollTimeline interface.
  assert_equals(animations[0].timeline, null,
                'scroll animation should not return scroll timeline');
}, 'Document.getAnimation() should include scroll animations');

</script>
</body>
